Added Matrix Class Moved Vector Sum to partial definition VectorBase now in Datascience.Core namespace AccessVal -> GetValue VectorBase ToCSV performance improved
